Rename getProblem → singleProblemGetter
