+++ src/api/CMakeLists.txt
@@ -4,7 +4,7 @@ add_library(Antares::solver_api ALIAS solver_api)
 set(PUBLIC_HEADERS
         include/antares/api/SimulationResults.h
         include/antares/api/solver.h
-	include/antares/api/getProblem.h
+	include/antares/api/singleProblemGetter.h
 )
 
 set(PRIVATE_HEADERS
@@ -18,7 +18,7 @@ target_sources(solver_api
         API.cpp
         SimulationObserver.cpp
         SimulationResults.cpp
-	getProblem.cpp
+	singleProblemGetter.cpp
         ${PUBLIC_HEADERS}
         ${PRIVATE_HEADERS}
 )
